# [reland][aotinductor] Add example_value metadata to nodes